requirements.txt
ExifRead == 2.1.1
path.py == 7.3
python_dateutil == 2.4.2
path == 0.33
